+++ doc/release_notes.md
@@ -28,6 +28,8 @@ commands which show an explanation about the command.
 - PKSMacroC supports now [] operators to extract single characters or a range of characters from a string or one or more words from a string array.
   or in assigments to change elements of an array.
 - PKSMacroC supports now global (namespace local / static) variables. Namespaces are currently limited however.
+- Long running macroC macros can now be interrupted - a progress bar pops up after a while to cancel them (e.g. if the macro code contains
+  an error such as an endless loop).
 - One can now debug print the low level instructions of PKSMacroC macros (from the macro dialog).
 - /* */ comments in macros are now supported and automatically attached to a macro as a comment (e.g. used as help in code completion), 
   if placed in front of a macro definition.